api_banco_dado/serializers/contato_serializer.py
from rest_framework import serializers
from core.models.Contato import Contato

class ContatoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contato
        fields = ('__all__')