requirements.txt
astroid==2.0.1
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.7
django-widget-tweaks==1.4.3
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.6.0
Pillow==5.3.0
psycopg2==2.7.5
pylint==2.0.0
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
djangorestframework==3.9.2
requests